requirements.txt
fastapi
uvicorn
pandas
openai
python-multipart
python-dotenv
